Added strobe light effect warning
Letting Chingu's know not to use any animation that is strobe like, or produces too much flashing imagery, and warning those who may be affected by strobe that this is possible when writing and testing a game code.  Advise them to take extreme caution if they are sensitive to these effects.